Trait pallet_transaction_payment_rpc::TransactionPaymentApiServer

pub trait TransactionPaymentApiServer<BlockHash, ResponseType>: Sized + Send + Sync + 'static {
    // Required methods
    fn query_info(
        &self,
        encoded_xt: Bytes,
        at: Option<BlockHash>
    ) -> RpcResult<ResponseType>;
    fn query_fee_details(
        &self,
        encoded_xt: Bytes,
        at: Option<BlockHash>
    ) -> RpcResult<FeeDetails<NumberOrHex>>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self>
       where BlockHash: Send + Sync + 'static + DeserializeOwned,
             ResponseType: Send + Sync + 'static + Serialize { ... }
}

Server trait implementation for the TransactionPaymentApi RPC API.

Required Methods

fn query_info( &self, encoded_xt: Bytes, at: Option<BlockHash> ) -> RpcResult<ResponseType>

fn query_fee_details( &self, encoded_xt: Bytes, at: Option<BlockHash> ) -> RpcResult<FeeDetails<NumberOrHex>>

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where BlockHash: Send + Sync + 'static + DeserializeOwned, ResponseType: Send + Sync + 'static + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.